# LegalConnect.API/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "LegalConnect API",
    "description": "Lawyer Marketplace Platform API",
    "version": "v1"
  },
  "paths": {
...
    "/api/cases/{caseId}/activities": {
      "get": {
        "tags": [
          "CaseActivities"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "caseId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "CaseActivities"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "caseId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AddCaseActivityDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/cases/{caseId}/documents": {
      "get": {
        "tags": [
          "CaseDocuments"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "caseId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "DocumentType",
            "type": "string"
          },
          {
            "in": "query",
            "name": "IsPrivate",
            "type": "boolean"
          },
          {
            "in": "query",
            "name": "PageNumber",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "PageSize",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "CaseDocuments"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "caseId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "file",
            "type": "file"
          },
          {
            "in": "formData",
            "name": "DocumentType",
            "required": true,
            "type": "string",
            "maxLength": 100
          },
          {
            "in": "formData",
            "name": "IsPrivate",
            "type": "boolean"
          },
          {
            "in": "formData",
            "name": "SharedWithLawyerIds",
            "type": "array",
            "items": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/case-documents/{id}/download": {
      "get": {
        "tags": [
          "CaseDocuments"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/case-documents/{id}": {
      "delete": {
        "tags": [
          "CaseDocuments"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/cases": {
      "get": {
        "tags": [
          "Cases"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "SearchTerm",
            "type": "string"
          },
          {
            "in": "query",
            "name": "Status"
          },
          {
            "in": "query",
            "name": "CaseType",
            "type": "string"
          },
          {
            "in": "query",
            "name": "Court",
            "type": "string"
          },
          {
            "in": "query",
            "name": "ClientProfileId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "FilingDateFrom",
            "type": "string",
            "format": "date-time"
          },
          {
            "in": "query",
            "name": "FilingDateTo",
            "type": "string",
            "format": "date-time"
          },
          {
            "in": "query",
            "name": "SortBy",
            "type": "string"
          },
          {
            "in": "query",
            "name": "SortDescending",
            "type": "boolean"
          },
          {
            "in": "query",
            "name": "PageNumber",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "PageSize",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Cases"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateCaseDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/cases/{id}": {
      "get": {
        "tags": [
          "Cases"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Cases"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateCaseDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Cases"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/cases/{id}/status": {
      "put": {
        "tags": [
          "Cases"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateStatusDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/cases/{id}/close": {
      "put": {
        "tags": [
          "Cases"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CloseCaseDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/cases/{id}/lawyers": {
      "get": {
        "tags": [
          "Cases"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Cases"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AddCaseLawyerDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/cases/{id}/lawyers/{lawyerProfileId}": {
      "delete": {
        "tags": [
          "Cases"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "lawyerProfileId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "definitions": {
    "ActivityType": {
      "format": "int32",
      "enum": [
        0,
        1,
        2,
        3,
        4,
        5,
        6
      ],
      "type": "integer"
    },
    "AddCaseActivityDto": {
      "required": [
        "activityType",
        "description"
      ],
      "type": "object",
      "properties": {
        "activityType": {
          "$ref": "#/definitions/ActivityType"
        },
        "description": {
          "maxLength": 1000,
          "minLength": 1,
          "type": "string"
        },
        "activityDate": {
          "format": "date-time",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "AddCaseLawyerDto": {
      "type": "object",
      "properties": {
        "lawyerProfileId": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
...
    "CaseStatus": {
      "format": "int32",
      "enum": [
        0,
        1,
        2,
        3
      ],
      "type": "integer"
    },
    "CloseCaseDto": {
      "required": [
        "outcome"
      ],
      "type": "object",
      "properties": {
        "outcome": {
          "maxLength": 2000,
          "minLength": 1,
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "CreateCaseDto": {
      "required": [
        "caseTitle",
        "caseType",
        "court",
        "description"
      ],
      "type": "object",
      "properties": {
        "caseTitle": {
          "maxLength": 300,
          "minLength": 1,
          "type": "string"
        },
        "caseNumber": {
          "maxLength": 100,
          "type": "string"
        },
        "caseType": {
          "maxLength": 100,
          "minLength": 1,
          "type": "string"
        },
        "court": {
          "maxLength": 200,
          "minLength": 1,
          "type": "string"
        },
        "clientProfileId": {
          "format": "int32",
          "type": "integer"
        },
        "lawyerProfileIds": {
          "type": "array",
          "items": {
            "format": "int32",
            "type": "integer"
          }
        },
        "appointmentId": {
          "format": "int32",
          "type": "integer"
        },
        "lawyerClientId": {
          "format": "int32",
          "type": "integer"
        },
        "filingDate": {
          "format": "date-time",
          "type": "string"
        },
        "nextHearingDate": {
          "format": "date-time",
          "type": "string"
        },
        "description": {
          "maxLength": 3000,
          "minLength": 1,
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "UpdateCaseDto": {
      "required": [
        "caseTitle",
        "caseType",
        "court",
        "description"
      ],
      "type": "object",
      "properties": {
        "caseTitle": {
          "maxLength": 300,
          "minLength": 1,
          "type": "string"
        },
        "caseNumber": {
          "maxLength": 100,
          "type": "string"
        },
        "caseType": {
          "maxLength": 100,
          "minLength": 1,
          "type": "string"
        },
        "court": {
          "maxLength": 200,
          "minLength": 1,
          "type": "string"
        },
        "filingDate": {
          "format": "date-time",
          "type": "string"
        },
        "nextHearingDate": {
          "format": "date-time",
          "type": "string"
        },
        "status": {
          "$ref": "#/definitions/CaseStatus"
        },
        "description": {
          "maxLength": 3000,
          "minLength": 1,
          "type": "string"
        },
        "outcome": {
          "maxLength": 2000,
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "UpdateStatusDto": {
      "type": "object",
      "properties": {
        "status": {
          "$ref": "#/definitions/CaseStatus"
        }
      },
      "additionalProperties": false
    }
  },
  "securityDefinitions": {
    "Bearer": { }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}